refactor: Make `citric.types` a private module
